+++ R/methods.R
@@ -1,9 +1,62 @@
+#' The propr Package
+#'
+#' @description
+#' Welcome to the \code{propr} package!
+#'
+#' To learn more about how to calculate proportionality, see
+#'  \code{\link{proportionality}}.
+#'
+#' To learn more about proportionality plots, see
+#'  \code{\link{visualize}}.
+#'
+#' To learn more about differential proportionality, see
+#'  \code{\link{prop2prob}}.
+#'
+#' To learn more about compositional data analysis, and its relevance
+#'  to biological count data, see the bundled vignette.
+#'
+#' @slot counts A matrix. Stores the original "count matrix" input.
+#' @slot logratio A matrix. Stores the log-ratio transformed "count matrix".
+#' @slot matrix A matrix. Stores the proportionality matrix calculated by
+#'  \code{phiRcpp} or \code{rhoRcpp}.
+#' @slot pairs A vector. Indexes the proportionality metrics of interest.
+#'
+#' @param object,x An object of class \code{propr}.
+#' @param subset Subsets via \code{object@counts[subset, ]}.
+#'  Use this argument to rearrange subject order.
+#' @param select Subsets via \code{object@counts[, select]}.
+#'  Use this argument to rearrange feature order.
+#' @param i Operation used for the subset indexing. Select from
+#'  "==", "=", ">", ">=", "<", "<=", "!=", or "all".
+#' @param j Reference used for the subset indexing. Provide a numeric
+#'  value to which to compare the proportionality measures in the
+#'  \code{@@matrix} slot.
+#' @param tiny A logical scalar. Toggles whether to pass the indexed
+#'  result through \code{\link{simplify}}.
+#' @param y Missing. Ignore. Leftover from the generic method
+#'  definition.
+#' @inheritParams visualize
+#'
+#' @name propr
+#' @useDynLib propr
+#' @importFrom methods show new
+#' @importFrom Rcpp sourceCpp
+NULL
+
+#' @rdname propr
+#' @export
+setClass("propr",
+         slots = c(
+           counts = "matrix",
+           logratio = "matrix",
+           matrix = "matrix",
+           pairs = "numeric"
+         )
+)
+
 #' @rdname propr
 #' @section Methods (by generic):
 #' \code{show:} Method to show \code{propr} object.
-#'
-#' @param object,x An object of class \code{propr}.
-#' @importFrom methods show
 #' @export
 setMethod("show", "propr",
           function(object){
@@ -31,12 +84,6 @@ setMethod("show", "propr",
 #' @rdname propr
 #' @section Methods (by generic):
 #' \code{subset:} Method to subset \code{propr} object.
-#'
-# #' @param x An object of class \code{propr}.
-#' @param subset Subsets via \code{object@counts[subset, ]}.
-#'  Use this argument to rearrange subject order.
-#' @param select Subsets via \code{object@counts[, select]}.
-#'  Use this argument to rearrange feature order.
 #' @export
 setMethod("subset", signature(x = "propr"),
           function(x, subset, select){
@@ -55,7 +102,7 @@ setMethod("subset", signature(x = "propr"),
 
             if(length(x@pairs) > 0){
 
-              cat("Alert: User must repopulate @pairs slot after `subset`.\n")
+              message("Alert: User must repopulate @pairs slot after `subset`.")
               x@pairs <- vector("numeric")
             }
 
@@ -66,17 +113,11 @@ setMethod("subset", signature(x = "propr"),
 #' @rdname propr
 #' @section Methods (by generic):
 #' \code{[:} Method to subset \code{propr} object.
-#'
-# #' @param x An object of class \code{propr}.
-#' @param i Operation used for the subset indexing. Select from
-#'  "==", "=", ">", ">=", "<", "<=", "!=", or "all".
-#' @param j Reference used for the subset indexing. Provide a numeric
-#'  value to which to compare the proportionality metrics.
 #' @aliases [,propr-method
 #' @docType methods
 #' @export
 setMethod('[', signature(x = "propr", i = "ANY", j = "ANY"),
-          function(x, i = "all", j){
+          function(x, i = "all", j, tiny = FALSE){
 
             if(i == "all"){
 
@@ -94,161 +135,88 @@ setMethod('[', signature(x = "propr", i = "ANY", j = "ANY"),
               stop("Reference not found. Index using e.g., `prop[\">\", .95]`.")
             }
 
-            x@pairs <- indexPairs(x@matrix, i, j)
+            newPairs <- indexPairs(x@matrix, i, j)
+            if(length(newPairs) == 0) message("Alert: Method failed to index any pairs.")
+            if(length(x@pairs) > 0) message("Alert: Appending prior index.")
+            x@pairs <- sort(union(x@pairs, newPairs))
 
-            if(length(x@pairs) == 0){
+            if(tiny){
 
-              stop("Method failed to index any pairs.")
-            }
-
-            return(x)
-          }
-)
-
-#' @rdname propr
-#' @section Methods (by generic):
-#' \code{plot:} Method to plot \code{propr} object.
-#'
-# #' @param x An object of class \code{propr}.
-#' @param y Missing. Ignore. Leftover from the generic method definition.
-#' @param title A character string. A title for the \code{propr} plot.
-#' @export
-setMethod("plot", signature(x = "propr", y = "missing"),
-          function(x, y, title = "Pairwise Proportionality"){
-
-            if(!requireNamespace("ggplot2", quietly = TRUE)){
-              stop("Uh oh! This plot method depends on ggplot2! ",
-                   "Try running: install.packages('ggplot2')")
-            }
-
-            if(!requireNamespace("ggthemes", quietly = TRUE)){
-              stop("Uh oh! This plot method depends on ggthemes! ",
-                   "Try running: install.packages('ggthemes')")
-            }
-
-            if(length(x@pairs) == 0){
-
-              cat("Alert: Generating plot using all feature pairs.\n")
-              V <- indexPairs(x@matrix, "all")
-              coord <- indexToCoord(V, nrow(x@matrix))
+              return(simplify(x))
 
             }else{
 
-              cat("Alert: Generating plot using indexed feature pairs.\n")
-              V <- x@pairs
-              coord <- indexToCoord(V, nrow(x@matrix))
+              return(x)
             }
-
-            # Melt *lr counts by feature pairs
-            nsubj <- nrow(x@logratio)
-            feat1 <- vector("numeric", length(V) * nsubj)
-            feat2 <- vector("numeric", length(V) * nsubj)
-            group <- vector("numeric", length(V) * nsubj)
-            for(i in 1:length(V)){
-
-              cat("Shaping pair", i, "...")
-              i.order <- order(x@logratio[, coord$feat1[i]])
-              feat1[((i-1)*nsubj + 1):((i-1)*nsubj + nsubj)] <- x@logratio[, coord$feat1[i]][i.order]
-              feat2[((i-1)*nsubj + 1):((i-1)*nsubj + nsubj)] <- x@logratio[, coord$feat2[i]][i.order]
-              group[((i-1)*nsubj + 1):((i-1)*nsubj + nsubj)] <- i
-            }
-
-            # Plot *lr-Y by *lr-X
-            cat("\n")
-            df <- data.frame("x.val" = feat1, "y.val" = feat2, "group" = group)
-            p <- ggplot2::ggplot(data = df,
-                                 ggplot2::aes_string(x = "x.val",
-                                                     y = "y.val",
-                                                     group = "group")) +
-              ggplot2::geom_path(ggplot2::aes(colour = factor(df$group))) +
-              ggplot2::labs(x = "*lr-transformed Abundance[1]",
-                            y = "*lr-transformed Abundance[2]") +
-              ggplot2::coord_equal(ratio = 1) +
-              ggthemes::theme_base() +
-              ggplot2::theme(legend.position = "none") +
-              ggplot2::ggtitle(title)
-            plot(p)
-
-            return(p)
           }
 )
 
 #' @rdname propr
 #' @section Methods (by generic):
 #' \code{plot:} Method to plot \code{propr} object.
-#'
-# #' @param object An object of class \code{propr}.
-# #' @param title A character string. A title for the \code{propr} plot.
-#' @param group A character or numeric vector. Supply feature groups for coloring.
-#'  Feature groups expected in the order they appear in \code{@@counts}.
-#' @importFrom stats as.dist as.dendrogram order.dendrogram
-#' @importFrom grDevices rainbow
 #' @export
-dendrogram <- function(object, title = "Proportional Clusters", group){
+setMethod("plot", signature(x = "propr", y = "missing"),
+          function(x, y, prompt = TRUE, plotly = FALSE){
 
-  if(suppressWarnings(!requireNamespace("fastcluster", quietly = TRUE))){
-    stop("Uh oh! This plot method depends on fastcluster! ",
-         "Try running: install.packages('fastcluster')")
+            smear(x, prompt = prompt, plotly = plotly)
+          }
+)
+
+#' @rdname propr
+#' @section Functions:
+#' \code{simplify:}
+#'  This convenience function takes an indexed \code{propr} object
+#'  and subsets the object based on that index. Then, it populates the
+#'  \code{@@pairs} slot of the new object with an updated version
+#'  of the original index. You can call \code{simplify} from within the
+#'  \code{[} method using the argument \code{tiny}.
+#' @export
+simplify <- function(object){
+
+  if(!class(object) == "propr" | length(object@pairs) == 0){
+
+    stop("Uh oh! This function requires an indexed 'propr' object.")
   }
 
-  if(suppressWarnings(!requireNamespace("dendextend", quietly = TRUE))){
-    stop("Uh oh! This plot method depends on dendextend! ",
-         "Try running: install.packages('dendextend')")
+  # Subset propr object based on index
+  coords <- indexToCoord(object@pairs, nrow(object@matrix))
+  selection <- sort(union(coords[[1]], coords[[2]]))
+  object@pairs <- vector("numeric")
+  new <- subset(object, select = selection)
+
+  # Repopulate the pairs slot
+  for(i in 1:length(coords[[1]])){
+
+    coords[[1]][i] <- which(selection == coords[[1]][i])
+    coords[[2]][i] <- which(selection == coords[[2]][i])
   }
 
-  if(length(object@pairs) == 0){
+  new@pairs <- (coords[[2]] - 1) * nrow(new@matrix) + (coords[[1]] - 1) + 1
 
-    cat("Alert: Generating plot using all feature pairs.\n")
-    i.feat <- 1:nrow(object@matrix)
+  return(new)
+}
 
-  }else{
+#' @rdname propr
+#' @section Functions:
+#' \code{adjacent:}
+#'  This function uses pairs indexed in the \code{@@pairs}
+#'  slot to build a symmetric adjacency matrix.
+#' @export
+adjacent <- function(object){
 
-    cat("Alert: Generating plot using indexed feature pairs.\n")
-    V <- object@pairs
-    coord <- indexToCoord(V, nrow(object@matrix))
-    i.feat <- sort(union(coord[[1]], coord[[2]]))
+  if(!class(object) == "propr" | length(object@pairs) == 0){
+
+    stop("Uh oh! This function requires an indexed 'propr' object.")
   }
 
-  # Align features with groups in data.frame
-  if(missing(group)) group <- 1
-  featnames <- colnames(object@logratio)[i.feat]
-  if(is.null(featnames)){
-    featnames <- paste("Feature", i.feat)
-  }
-  colorKey <- data.frame("feature" = featnames,  "group" = group, "color" = NA,
-                         stringsAsFactors = FALSE)
+  N <- nrow(object@matrix)
+  mat <- matrix(0, N, N)
+  mat[object@pairs] <- 1
+  diag(mat) <- 1
+  symRcpp(mat)
 
-  # Assign 'n' colors based on 'n' groups
-  grps <- unique(colorKey$group)
-  colors <- rainbow(length(grps))
-  for(i in 1:length(grps)){
-    colorKey[colorKey$group == grps[i], "color"] <- colors[i]
-  }
+  adj <- object
+  adj@matrix <- mat
 
-  # Build tree and color branches
-  if(object@matrix[1, 1] == 0){
-
-    # Convert phi into dist matrix
-    dist <- as.dist(object@matrix[i.feat, i.feat])
-    attr(dist, "Labels") <- featnames
-
-  }else if(object@matrix[1, 1] == 1){
-
-    # Convert rho into dist matrix
-    # See reference: http://research.stowers-institute.org/
-    #  mcm/efg/R/Visualization/cor-cluster/index.htm
-    dist <- as.dist(1 - abs(object@matrix[i.feat, i.feat]))
-    attr(dist, "Labels") <- featnames
-
-  }else{
-
-    stop("Matrix style not recognized.")
-  }
-
-  hc <- fastcluster::hclust(dist)
-  dend <- as.dendrogram(hc)
-  dendextend::labels_colors(dend) <- colorKey$color[order.dendrogram(dend)]
-  plot(dend, main = title)
-
-  return(hc)
+  return(adj)
 }